Refactor training utilities into separate io and utils modules
